[fix]: Rename dbFiles to camel case
